ViewModel logic change to business class
