+++ TTSToVideo.Business/TTSToVideo.Business.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\NetXP\NetXP\NetXP.csproj" />
+    <ProjectReference Include="..\TTSToVideo.Helpers\TTSToVideo.Helpers.csproj" />
+  </ItemGroup>
+
+</Project>
