+++ TTSToVideo.Helpers/Constants.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TTSToVideo.WPF.Helpers
+namespace TTSToVideo.Helpers
 {
-    internal static class Constants
+    public static class Constants
     {
         public const int MAX_PATH = 255 - 4 - 50;//-4 Extension
     }
