+++ TTSToVideo.Helpers/Implementations/ProgressBar.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TTSToVideo.Helpers.Implementations
+{
+    public class ProgressBar : IProgressBar
+    {
+        public void ShowMessage(string message)
+        {
+        }
+    }
+}
