+++ TTSToVideo.WPF/AssemblyInfo.cs
@@ -7,4 +7,5 @@ using System.Windows;
     ResourceDictionaryLocation.SourceAssembly //where the generic resource dictionary is located
                                               //(used if a resource is not found in the page,
                                               // app, or any theme specific resource dictionaries)
-)]
+)] 
+
