--- TTSToVideo.WPF/IMainWindow.cs
@@ -1,11 +1,0 @@
-﻿using TTSToVideo.WPF.ViewModel;
-
-namespace TTSToVideo
-{
-    internal interface IMainWindow
-    {
-        IVMMainWindow ViewModel { get; }
-
-        void Show();
-    }
-}
