+++ TTSToVideo.WPF/Models/Message.cs
@@ -1,0 +1,8 @@
+﻿namespace TTSToVideo.WPF.Models
+{
+    internal class Message
+    {
+        public string? Text { get; internal set; }
+        public MessageType Type { get; internal set; }
+    }
+}
