+++ TTSToVideo.WPF/Models/StatementImageModel.cs
@@ -1,0 +1,7 @@
+﻿namespace TTSToVideo.WPF.Models
+{
+    public class StatementImageModel
+    {
+        public string Path { get; internal set; }
+    }
+}
