+++ TTSToVideo.WPF/Pages/ConfigurationPage.xaml
@@ -3,7 +3,8 @@
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:TTSToVideo.WPF.Pages" xmlns:implementations="clr-namespace:TTSToVideo.WPF.ViewModel.Implementations" d:DataContext="{d:DesignInstance Type=implementations:VMConfiguration}"
+      xmlns:local="clr-namespace:TTSToVideo.WPF.Pages" xmlns:implementations="clr-namespace:TTSToVideo.WPF.ViewsModels" 
+      d:DataContext="{d:DesignInstance Type=implementations:VMConfiguration}"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
       Title="ConfigurationPage">
@@ -11,6 +12,8 @@
         <StackPanel>
             <Label>Music Dir</Label>
             <TextBox Text="{Binding Model.MusicDir, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
+            <Label>Global Negative Prompt</Label>
+            <TextBox Text="{Binding Model.NegativePrompt, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
         </StackPanel>
     </Grid>
 </Page>
