+++ TTSToVideo.WPF/Pages/TTSToVideoPage.xaml
@@ -4,9 +4,10 @@
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:TTSToVideo.WPF" xmlns:implementations="clr-namespace:TTSToVideo.WPF.ViewModel.Implementations" d:DataContext="{d:DesignInstance Type=implementations:VMTTSToVideoPage}"
+      xmlns:local="clr-namespace:TTSToVideo.WPF" xmlns:implementations="clr-namespace:TTSToVideo.WPF.ViewsModels"
+      d:DataContext="{d:DesignInstance Type=implementations:VMTTSToVideoPage}"
       mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="967"
+      d:DesignHeight="450" d:DesignWidth="1179"
       Background="White"
       Title="MainPage"
       >
@@ -71,12 +72,14 @@
                 <Label>Portrait Prompt</Label>
                 <TextBox 
                      Style="{ StaticResource ShortInlineTextbox}"
+                     IsEnabled="{Binding ElementName=EnablePortrait, Path=IsChecked}"
                     Text="{Binding Model.PortraitPrompt}"></TextBox>
                 <Label>Portrait Image</Label>
                 <TextBox  
                      TextWrapping="Wrap" 
                      AcceptsReturn="True" 
                      MinWidth="200"
+                     IsEnabled="{Binding ElementName=EnablePortrait, Path=IsChecked}"
                      Style="{ StaticResource ShortInlineTextbox}"
                      Text="{Binding Model.PortraitImagePath, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
                 ></TextBox>
@@ -87,6 +90,11 @@
                 <Button  IsEnabled="{Binding ElementName=EnablePortrait, Path=IsChecked}" 
                          Content="Generate Portrait (IA)" 
                          Command="{Binding GeneratePortraitImageCommand}"  />
+
+                <Button  IsEnabled="{Binding ElementName=EnablePortrait, Path=IsChecked}" 
+                         Content="Generate Video (IA)" 
+                         Command="{Binding GeneratePortraitVideoCommand}"  />
+
 
             </StackPanel>
