--- TTSToVideo.WPF/ViewsModels/IVMConfiguration.cs
@@ -1,19 +1,0 @@
-﻿using CommunityToolkit.Mvvm.Input;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TTSToVideo.WPF.Models;
-
-namespace TTSToVideo.WPF.ViewModel
-{
-    public interface IVMConfiguration : INotifyPropertyChanged
-    {
-        ConfigurationModel Model { get; set; }
-        Task Init();
-        AsyncRelayCommand SaveCommand { get; set; }
-
-    }
-}
