--- TTSToVideo.WPF/ViewsModels/Implementations/Message.cs
@@ -1,8 +1,0 @@
-﻿namespace TTSToVideo.WPF.ViewModel.Implementations
-{
-    internal class Message
-    {
-        public string Text { get; internal set; }
-        public MessageType Type { get; internal set; }
-    }
-}
