+++ TTSToVideo.sln
@@ -15,6 +15,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetXP.TTSs.ElevenLabs", "..
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Framework", "Framework", "{FB61784D-3AE8-49AB-B7CE-34FAF7AD023B}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TTSToVideo.Helpers", "TTSToVideo.Helpers\TTSToVideo.Helpers.csproj", "{1DAD8988-DA0B-4681-9462-7354588BEA97}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetXP.UnitTest", "..\..\NetXP\NetXP.NetCore.UnitTest\NetXP.UnitTest.csproj", "{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TTSToVideo.UnitTests", "TTSToVideo.UnitTests\TTSToVideo.UnitTests.csproj", "{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TTSToVideo.Business", "TTSToVideo.Business\TTSToVideo.Business.csproj", "{FB89FA0C-7F3B-4111-9B95-200651F0211A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -63,6 +71,38 @@ Global
 		{BC679B08-1D03-4202-A03B-2BD53D8FB584}.Staging|Any CPU.Build.0 = Debug|Any CPU
 		{BC679B08-1D03-4202-A03B-2BD53D8FB584}.Testing|Any CPU.ActiveCfg = Debug|Any CPU
 		{BC679B08-1D03-4202-A03B-2BD53D8FB584}.Testing|Any CPU.Build.0 = Debug|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Staging|Any CPU.ActiveCfg = Debug|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Staging|Any CPU.Build.0 = Debug|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Testing|Any CPU.ActiveCfg = Debug|Any CPU
+		{1DAD8988-DA0B-4681-9462-7354588BEA97}.Testing|Any CPU.Build.0 = Debug|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Staging|Any CPU.ActiveCfg = Staging|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Staging|Any CPU.Build.0 = Staging|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Testing|Any CPU.ActiveCfg = Debug|Any CPU
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5}.Testing|Any CPU.Build.0 = Debug|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Staging|Any CPU.ActiveCfg = Debug|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Staging|Any CPU.Build.0 = Debug|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Testing|Any CPU.ActiveCfg = Debug|Any CPU
+		{20EA46F1-A08C-4142-B3A3-C165A0D2BB9E}.Testing|Any CPU.Build.0 = Debug|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Staging|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Staging|Any CPU.Build.0 = Debug|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Testing|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB89FA0C-7F3B-4111-9B95-200651F0211A}.Testing|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -72,6 +112,7 @@ Global
 		{DBD63AB4-036D-437C-AD1E-4345A00A7471} = {FB61784D-3AE8-49AB-B7CE-34FAF7AD023B}
 		{E77C443E-3E86-4963-ABBE-9A3B65462DED} = {FB61784D-3AE8-49AB-B7CE-34FAF7AD023B}
 		{BC679B08-1D03-4202-A03B-2BD53D8FB584} = {FB61784D-3AE8-49AB-B7CE-34FAF7AD023B}
+		{A32FB66E-D017-4D98-96BD-B4C8849ED0F5} = {FB61784D-3AE8-49AB-B7CE-34FAF7AD023B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BF645A19-807E-4D55-A008-672D6E69800A}
